backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests>=2.32.2
gunicorn==21.2.0
fpdf2==2.7.6
matplotlib>=3.10.0 